Prepare working directory on server startup
before binding to socket. Instead of at the time the working directory
is needed.

This allows us to remove `withMkTempWorkDir` function and by extension
simplify the `handleConnection` and get better behavior as the problem
to create working directory manifests immediately when starting the
server.

Case where the working directory disappears during runtime of the server
is unhandled as it is deemed unlikely. It will show up as an IOError
during the handleConnection.